update reinonlein skills

+++ apps/alcoholvrijheid/lib/models/alcoholvrijerd.dart
@@ -1,7 +1,7 @@
 class Alcoholvrijerd {
   final String name;
   final DateTime stopdate;
-  final double geld;
+  final int geld;
   final int bier;
   final int wijn;
   final int sterk;
